pyproject.toml
[project]
name = "weather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anthropic==0.44.0",
    "anyio==4.8.0",
    "certifi==2024.12.14",
    "click==8.1.8",
    "colorama==0.4.6",
    "distro==1.9.0",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "idna==3.10",
    "jiter==0.8.2",
    "markdown-it-py==3.0.0",
    "mcp[cli]==1.2.0",
    "mdurl==0.1.2",
    "pydantic==2.10.5",
    "pydantic-core==2.27.2",
    "pydantic-settings==2.7.1",
    "pygments==2.19.1",
    "python-dotenv==1.0.1",
    "rich==13.9.4",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "sse-starlette==2.2.1",
    "starlette==0.45.2",
    "typer==0.15.1",
    "typing-extensions==4.12.2",
    "uvicorn==0.34.0",
]
